requirements.txt
pymongo
sshtunnel
yfinance
pandas
pandas_market_calendars
datetime
openmeteo-requests
requests-cache
retry-requests
python-dotenv
cryptography
snowflake-connector-python[pandas]
pyarrow